A remote procedure calling has been implemented into order api service

+++ OrderService/OrderService.Application/MediatR/Order/Handlers/DeleteOrderHandler.cs
@@ -1,9 +1,11 @@
 ﻿using Hangfire;
 using MediatR;
+using Microsoft.AspNetCore.Http;
 using OrderService.Application.DTOs.Order;
 using OrderService.Application.Interfaces.Repositories.NoSql;
 using OrderService.Application.Interfaces.Repositories.Sql;
 using OrderService.Application.MediatR.Order.Commands;
+using OrderService.Application.TokenParsers.Interfaces;
 using OrderService.Domain.Exceptions;
 
 namespace OrderService.Application.MediatR.Order.Handlers
@@ -13,14 +15,20 @@ namespace OrderService.Application.MediatR.Order.Handlers
         private readonly ISqlOrderRepository _sqlOrderRepository;
         private readonly INoSqlOrderRepository _noSqlOrderRepository;
         private readonly IBackgroundJobClient _backgroundJobClient;
+        private readonly ITokenParser _tokenParser;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
         public DeleteOrderHandler(ISqlOrderRepository sqlClientRepository,
             INoSqlOrderRepository noSqlClientRepository,
-            IBackgroundJobClient backgroundJobClient)
+            IBackgroundJobClient backgroundJobClient,
+            ITokenParser tokenParser,
+            IHttpContextAccessor httpContextAccessor)
         {
             _sqlOrderRepository = sqlClientRepository;
             _noSqlOrderRepository = noSqlClientRepository;
             _backgroundJobClient = backgroundJobClient;
+            _tokenParser = tokenParser;
+            _httpContextAccessor = httpContextAccessor;
         }
 
         public async Task Handle(DeleteOrderCommand request,
@@ -33,6 +41,15 @@ namespace OrderService.Application.MediatR.Order.Handlers
             {
                 throw new NotFoundException(nameof(Domain.Entities.Order),
                     request.Id.ToString(), typeof(Domain.Entities.Order));
+            }
+
+            int subjectId = _tokenParser
+                .ParseSubjectId(_httpContextAccessor?.HttpContext?.Request.Headers);
+
+            if (orderDTO.ReadClientDTO.Id != subjectId)
+            {
+                throw new AuthorizationException(orderDTO.ReadTableDTO.Id.ToString(),
+                    ExceptionMessages.NotClientBookingMessage);
             }
 
             await _sqlOrderRepository.DeleteAsync(request.Id, cancellationToken);
